Update WebRTC uninstall, add Teams Meeting Add-in install and uninstall
The product code for the WebRTC service changed after a recent update. By querying for the package name and grabbing the product code from that, we are able to always find the WebRTC product code, even if Microsoft changes it in the future.

This update will also check for the Teams Meeting add-in and uninstall it if it is present, then reinstall the latest version. The meeting add-in is no longer bundled by default in the Teams installer, and must instead be installed manually from the MSIX package.